matrix_bot_api: Added dynamic handler adding/removing + arguments
Dynamic handler management:
The bot API is now able to dynamically remove registered handlers
_without_ restarting. This allows e.g. the registration/deletion of new
bot features/handlers through a matrix bot command.

Additional Handler Argument:
All handlers now have an extra (3rd) argument besides (room, event).
This allows to pass information to the handler while registering. A
possible use case is a handler, which reacts in the same way on
different keywords, but needs to know about the context it was
registered in.